refactor(popup): colocate tab message wrappers, extract messaging helpers
